# openapi.yaml
openapi: 3.0.3
info:
  title: Gopher Holes Unlimited API
  description: |
    Gopher Holes Unlimited (GHU) is a fictional service to track gophers, holes,
    and our continuous quest to keep them from destroying our gardens.

    This version is powered by a single "mono-Lambda" that routes requests with
    an AWS Lambda Powertools event-handler controller. Writes are committed with
    DynamoDB transactions that also enqueue domain events (the transactional
    outbox pattern); those events are relayed to EventBridge and drive
    choreography such as automatically linking a new gopher to holes at its
    location.

    ![Gopher Holes Unlimited](https://readysetcloud.s3.amazonaws.com/GHU.png)
  version: 2.0.0

tags:
  - name: Gophers
    description: The little varmints themselves
  - name: Holes
    description: Details on the gopher holes

servers:
  - description: dev
    url: https://api.gopherholesunlimited.com

security:
  - api_key: []

paths:
  /gophers:
    post:
      summary: Track a new gopher
      description: |
        Adds a gopher to the registry. Creation is synchronous and returns the new
        id. Send an `Idempotency-Key` header to make retries safe.
      tags: [Gophers]
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGopher'
      responses:
        201:
          $ref: '#/components/responses/Created'
        400:
          $ref: '#/components/responses/BadRequest'
    get:
      summary: View a list of all gophers
      tags: [Gophers]
      responses:
        200:
          description: A list of gophers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gopher'

  /gophers/{gopherId}:
    parameters:
      - $ref: '#/components/parameters/gopherId'
    get:
      summary: Get the details of a specific gopher
      tags: [Gophers]
      parameters:
        - name: include
          in: query
          required: false
          description: Set to `holes` to include the gopher's linked holes.
          schema:
            type: string
            enum: [holes]
      responses:
        200:
          description: The requested gopher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gopher'
        404:
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update a subset of details of a specific gopher
      tags: [Gophers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGopher'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a gopher
      tags: [Gophers]
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        404:
          $ref: '#/components/responses/NotFound'

  /gophers/{gopherId}/statuses:
    parameters:
      - $ref: '#/components/parameters/gopherId'
    post:
      summary: Record a new status for a gopher
      description: Sets the current status of a gopher. All status history is retained.
      tags: [Gophers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [at large, trapped, deceased, unknown]
                  example: trapped
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /holes:
    post:
      summary: Track a new hole
      description: |
        Adds a hole to the system. If a `gopherId` is supplied, the hole is linked
        to that gopher asynchronously via a domain event. Send an `Idempotency-Key`
        header to make retries safe.
      tags: [Holes]
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHole'
      responses:
        201:
          $ref: '#/components/responses/Created'
        400:
          $ref: '#/components/responses/BadRequest'
    get:
      summary: View a list of all holes
      tags: [Holes]
      parameters:
        - name: status
          in: query
          required: false
          description: Filter holes by their current status
          schema:
            type: string
            enum: [filled, visible]
      responses:
        200:
          description: A list of holes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hole'

  /holes/{holeId}:
    parameters:
      - $ref: '#/components/parameters/holeId'
    get:
      summary: Get the details of a specific hole
      tags: [Holes]
      parameters:
        - name: include
          in: query
          required: false
          description: Set to `gopher` to include a summary of the linked gopher.
          schema:
            type: string
            enum: [gopher]
      responses:
        200:
          description: The requested hole
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hole'
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Replace the details of a specific hole
      tags: [Holes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHole'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update some of the details of a specific hole
      tags: [Holes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHole'
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /holes/{holeId}/statuses:
    parameters:
      - $ref: '#/components/parameters/holeId'
    post:
      summary: Update the status of a hole
      tags: [Holes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [filled, visible]
                  example: filled
      responses:
        204:
          $ref: '#/components/responses/NoContent'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header

  parameters:
    gopherId:
      name: gopherId
      in: path
      description: Unique identifier for a gopher
      required: true
      schema:
        type: string
        example: 01JBX8P0Q5N6E3S9K2WJ7YQZ4T
    holeId:
      name: holeId
      in: path
      description: Unique identifier for a hole
      required: true
      schema:
        type: string
        example: 01JBX8P4R2C7A1M8H6D0V5B3ZK
    idempotencyKey:
      name: Idempotency-Key
      in: header
      description: Optional client-supplied key that makes create requests safe to retry.
      required: false
      schema:
        type: string
        example: 6d3f2b7e-2b1a-4c9e-8f0a-1a2b3c4d5e6f

  schemas:
    Location:
      description: A physical location, expressed as coordinates or a street address.
      oneOf:
        - type: object
          required: [latitude, longitude]
          properties:
            latitude:
              type: string
              example: '33.067067'
            longitude:
              type: string
              example: '-96.801599'
        - type: object
          required: [city, state]
          properties:
            addressLine1:
              type: string
              example: 123 Main Street
            addressLine2:
              type: string
              example: 'Apt #2212'
            city:
              type: string
              example: Plano
            state:
              type: string
              example: TX

    NewGopher:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
          example: Gerry
        type:
          type: string
          enum: [Western Pocket, Eastern Pocket, Geomys]
        sex:
          type: string
          enum: [male, female, no preference, unknown]
        picture:
          type: string
          example: https://gopherhunter.s3.amazonaws.com/gerry.jpg
        status:
          type: string
          enum: [at large, trapped, deceased, unknown]
        color:
          type: string
          example: brown
        location:
          $ref: '#/components/schemas/Location'
        comment:
          type: string

    UpdateGopher:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Western Pocket, Eastern Pocket, Geomys]
        sex:
          type: string
          enum: [male, female, no preference, unknown]
        picture:
          type: string
        status:
          type: string
          enum: [at large, trapped, deceased, unknown]
        color:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        comment:
          type: string

    Gopher:
      allOf:
        - $ref: '#/components/schemas/NewGopher'
        - type: object
          properties:
            id:
              type: string
            timesSeen:
              type: integer
              example: 42
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            holes:
              type: array
              description: Present only when requested with `?include=holes`.
              items:
                $ref: '#/components/schemas/LinkedHole'

    LinkedHole:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [filled, visible]

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: The 'name' field is required and is missing.

    NewHole:
      type: object
      required: [location, description]
      properties:
        description:
          type: string
          maxLength: 100
          example: Near the fire hydrant
        gopherId:
          type: string
          description: The gopher that dug the hole
        location:
          $ref: '#/components/schemas/Location'
        status:
          type: string
          enum: [filled, visible]
        comment:
          type: string

    UpdateHole:
      type: object
      minProperties: 1
      properties:
        description:
          type: string
          maxLength: 100
        gopherId:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        status:
          type: string
          enum: [filled, visible]
        comment:
          type: string

    Hole:
      allOf:
        - $ref: '#/components/schemas/NewHole'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            gopher:
              type: object
              description: Present only when requested with `?include=gopher`.
              properties:
                id:
                  type: string
                name:
                  type: string

  responses:
    Created:
      description: The entity was created
      content:
        application/json:
          schema:
            type: object
            required: [id]
            properties:
              id:
                type: string
                example: 01JBX8P0Q5N6E3S9K2WJ7YQZ4T
    BadRequest:
      description: The request was invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The entity could not be found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoContent:
      description: The request succeeded and there is no content to return